feat(기능) 컨트롤러의 jwt토근 기능 추가
bookcontroller
cartcontroller
likecontroller
ordercontroller
usercontroller
-> jwt 토근 기능 추가

and

auth.js추가

+++ controller/CartController.js
@@ -1,14 +1,34 @@
 const conn = require("../mariadb");
+const ensureAuthorization = require("../auth");
 const { StatusCodes } = require("http-status-codes");
 const { use } = require("../routes/carts");
 const jwt = require("jsonwebtoken");
-const dotenv = require("dotenv");
-dotenv.config();
 
 const addToCart = (req, res) => {
   const { book_id, quantity } = req.body;
 
   let authorization = ensureAuthorization(req);
+
+  if (authorization instanceof jwt.TokenExpiredError) {
+    return res.status(StatusCodes.UNAUTHORIZED).json({
+      message: "로그인 세션이 만료되었습니다. 다시 로그인 하세요.",
+    });
+  } else if (authorization instanceof jwt.JsonWebTokenError) {
+    return res.status(StatusCodes.BAD_REQUEST).json({
+      message: "잘못된 토근입니다.",
+    });
+  } else {
+    let sql = `INSERT INTO cartItems (book_id, quantity, user_id) VALUES (?,?,?)`;
+    let values = [book_id, quantity, authorization.id];
+    conn.query(sql, values, (err, results) => {
+      if (err) {
+        console.log(err);
+        return res.status(StatusCodes.BAD_REQUEST).end();
+      }
+
+      return err;
+    });
+  }
 
   let sql = "INSERT INTO cartItems (book_id, quantity, user_id) VALUES(?,?,?)";
   let values = [book_id, quantity, authorization.id];
@@ -39,50 +59,50 @@ const getCartItems = (req, res) => {
     let sql = `select cartItems.id, book_id, title, summary, quantity, price 
         FROM cartItems 
         LEFT JOIN books ON cartItems.book_id = books.id 
-        WHERE user_id=? AND cartItems.id IN(?)`;
-    let values = [authorization.id, selected];
+        WHERE user_id=?`;
+
+    let values = [authorization.id];
+
+    if (selected) {
+      // 주문서 작성 시 '선택한 장바구니 목록 조회'
+      sql += ` AND cartItems.id IN (?)`;
+      values.push(selected);
+    }
+
     conn.query(sql, values, (err, results) => {
       if (err) {
         console.log(err);
         return res.status(StatusCodes.BAD_REQUEST).end();
       }
-
-      return err;
+      return res.status(StatusCodes.OK).json(results);
     });
   }
 };
 
 const removeCartItems = (req, res) => {
-  const cartItems = req.params.id;
+  let authorization = ensureAuthorization(req, res);
 
-  let sql = "DELETE FROM cartItems WHERE id = ?;";
-
-  conn.query(sql, cartItems, (err, results) => {
-    if (err) {
-      console.log(err);
-      return res.status(StatusCodes.BAD_REQUEST).end();
-    }
-    return res.status(StatusCodes.OK).json(results);
-  });
-};
-
-function ensureAuthorization(req, res) {
-  try {
-    let receivedJwt = req.headers["authorization"];
-    console.log("received jwt : ", receivedJwt);
-
-    let decodedJwt = jwt.verify(receivedJwt, process.env.PRIVATE_KEY);
-    console.log(decodedJwt);
-
-    return decodedJwt;
-  } catch (err) {
-    console.log(err.name);
-    console.log(err.message);
-
+  if (authorization instanceof jwt.TokenExpiredError) {
     return res.status(StatusCodes.UNAUTHORIZED).json({
       message: "로그인 세션이 만료되었습니다. 다시 로그인 하세요.",
     });
+  } else if (authorization instanceof jwt.JsonWebTokenError) {
+    return res.status(StatusCodes.BAD_REQUEST).json({
+      message: "잘못된 토근입니다.",
+    });
+  } else {
+    const cartItems = req.params.id;
+
+    let sql = "DELETE FROM cartItems WHERE id = ?;";
+
+    conn.query(sql, cartItems, (err, results) => {
+      if (err) {
+        console.log(err);
+        return res.status(StatusCodes.BAD_REQUEST).end();
+      }
+      return res.status(StatusCodes.OK).json(results);
+    });
   }
-}
+};
 
 module.exports = { addToCart, getCartItems, removeCartItems };
